requirements.txt
Flask==2.3.3
requests==2.31.0
google-generativeai==0.3.2
groq==0.4.1
aiohttp==3.9.1
asyncio==3.4.3
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0